requirements.txt
Flask
gunicorn
requests
together
python-dotenv